tytocam/requirements.txt
picamera==1.13
getmac==0.8.2
pynmcli==1.0.5
PyGObject==3.30.4
Django==2.2.4
numpy==1.16.2
requests==2.21.0
boto3==1.17.66
paho_mqtt==1.5.1
pyserial==3.5
